moved dtos to subdirectory
