+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/NamespacedKeyDTO.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.config;
+package org.braekpo1nt.mctmanager.config.dto;
 
 import com.google.common.base.Preconditions;
 import org.bukkit.NamespacedKey;
