+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/YawPitch.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.config;
+package org.braekpo1nt.mctmanager.config.dto;
 
 /**
  * Simple record to store the yaw and pitch of a Location. Useful for serializing the directional portion of a Location in JSON.
