+++ src/main/java/org/braekpo1nt/mctmanager/games/colossalcombat/config/ColossalCombatConfig.java
@@ -1,9 +1,9 @@
 package org.braekpo1nt.mctmanager.games.colossalcombat.config;
 
 import com.google.gson.JsonElement;
-import org.braekpo1nt.mctmanager.config.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.config.LocationDTO;
-import org.braekpo1nt.mctmanager.config.inventory.PlayerInventoryDTO;
+import org.braekpo1nt.mctmanager.config.dto.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.dto.LocationDTO;
+import org.braekpo1nt.mctmanager.config.dto.inventory.PlayerInventoryDTO;
 import org.jetbrains.annotations.Nullable;
 
 /**
