+++ src/main/java/org/braekpo1nt/mctmanager/games/game/mecha/config/MechaStorageUtil.java
@@ -7,7 +7,7 @@ import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.braekpo1nt.mctmanager.Main;
 import org.braekpo1nt.mctmanager.config.GameConfigStorageUtil;
-import org.braekpo1nt.mctmanager.config.YawPitch;
+import org.braekpo1nt.mctmanager.config.dto.YawPitch;
 import org.braekpo1nt.mctmanager.utils.EntityUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
