+++ src/main/java/org/braekpo1nt/mctmanager/hub/config/HubStorageUtil.java
@@ -3,7 +3,7 @@ package org.braekpo1nt.mctmanager.hub.config;
 import com.google.common.base.Preconditions;
 import org.braekpo1nt.mctmanager.Main;
 import org.braekpo1nt.mctmanager.config.GameConfigStorageUtil;
-import org.braekpo1nt.mctmanager.config.LocationDTO;
+import org.braekpo1nt.mctmanager.config.dto.LocationDTO;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
